+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next Release
 
+ * Refactor strand object, remove second argument
+
+## Next Release
+
  * Add `stream.writeMessage`
  * Add `strand.getPixel`
  * Add parser